update to gcc 4.6.2 and binutils 2.22
changed the strategy for matching libgccs a bit, it was getting out of hand
and somewhat impossible to keep making a seperate libgcc for each core active.

Now it requires a bit more manual setting of the fpu on the command line or
you'll get a bad match. So it goes.